graphql: move middleware code into separate file
